change imports and rename src folder
