+++ moddingway/bot.py
@@ -1,13 +1,13 @@
 import discord
 from discord.ext.commands import Bot
-from settings import get_settings
-from commands.exile_commands import create_exile_commands
-from commands.ban_commands import create_ban_commands
-from commands.strikes_command import create_strikes_commands
-from commands.helper import create_bot_errors
-from commands.slowmode_commands import create_slowmode_commands
+from moddingway.settings import get_settings
+from moddingway.commands.exile_commands import create_exile_commands
+from moddingway.commands.ban_commands import create_ban_commands
+from moddingway.commands.strikes_command import create_strikes_commands
+from moddingway.commands.helper import create_bot_errors
+from moddingway.commands.slowmode_commands import create_slowmode_commands
 import logging
-import workers
+from moddingway import workers
 
 settings = get_settings()
 logger = logging.getLogger(__name__)
