+++ moddingway/commands/ban_commands.py
@@ -1,8 +1,8 @@
 import discord
 from discord.ext.commands import Bot
-from settings import get_settings
-from services.ban_service import ban_user
-from util import is_user_moderator
+from moddingway.settings import get_settings
+from moddingway.services.ban_service import ban_user
+from moddingway.util import is_user_moderator
 from .helper import create_logging_embed, create_response_context
 
 settings = get_settings()
