+++ moddingway/constants.py
