+++ moddingway/database/strikes_database.py
